ECommerce/requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.0.5
django-allauth==0.51.0
django-bootstrap-v5==1.0.11
django-bootstrap5==22.1
django-configuration==0.0.3
django-configurations==2.3.2
django-crispy-forms==1.14.0
django-polymorphic==3.1.0
django-rename-app==0.1.3
django-reset-migrations==0.4.0
djangorestframework==3.13.1
entrypoints==0.4
executing==0.8.3
Flask==2.1.2
idna==3.3
ipykernel==6.15.0
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.4
jupyter-core==4.10.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mysqlclient==2.1.1
nest-asyncio==1.5.5
numpy==1.23.0
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.29
psutil==5.9.1
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.1
pywin32==304
pyzmq==23.2.0
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.3.0
tornado==6.1
traitlets==5.3.0
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
